random walk traces generation
