+++ python/dgl/contrib/sampling/__init__.py
@@ -1,1 +1,2 @@
 from .sampler import NeighborSampler
+from .randomwalk import *
