+++ python/dgl/graph_index.py
@@ -677,6 +677,20 @@ class GraphIndex(object):
         return [SubgraphIndex(rst(i), self, utils.toindex(rst(num_subgs + i)),
                               utils.toindex(rst(num_subgs * 2 + i))) for i in range(num_subgs)]
 
+    def random_walk(self, seeds, num_traces, num_hops):
+        """Random walk sampling.
+
+        Returns a user Tensor of random walk traces with shape
+        (num_seeds, num_traces, num_hops + 1)
+        """
+        if len(seeds) == 0:
+            return utils.toindex([])
+
+        seeds = seeds.todgltensor()
+        traces = _CAPI_DGLGraphRandomWalk(self._handle, seeds, num_traces, num_hops)
+
+        return F.zerocopy_from_dlpack(traces.to_dlpack())
+
     def to_networkx(self):
         """Convert to networkx graph.
